Crate substrate_consensus_common

Common utilities for building and using consensus engines in substrate.

Much of this crate is unstable and thus the API is likely to undergo change. Implementors of traits should not rely on the interfaces to remain the same.

Re-exports

pub use self::error::Error;

pub use block_import::BlockImport;

pub use block_import::BlockOrigin;

pub use block_import::ForkChoiceStrategy;

pub use block_import::ImportedAux;

pub use block_import::ImportBlock;

pub use block_import::ImportResult;

pub use block_import::JustificationImport;

pub use block_import::FinalityProofImport;

pub use block_import::FinalityProofRequestBuilder;

Modules

block_import
Block import helpers.

error
Error types in Consensus

evaluation
Block evaluation and evaluation errors.

import_queue
Import Queue primitive: something which can verify and import blocks.

offline_tracker
Tracks offline validators.

well_known_cache_keys
A list of all well known keys in the cache.

Structs

InherentData
Inherent data to include in a block.

NoNetwork
A synchronization oracle for when there is no network.

Traits

Environment
Environment producer for a Consensus instance. Creates proposer instance and communication streams.

Proposer
Logic for a proposer.

SelectChain
The SelectChain trait defines the strategy upon which the head is chosen if multiple forks are present for an opaque definition of "best" in the specific chain build.

SyncOracle
An oracle for when major synchronization work is being undertaken.