Struct substrate_consensus_common::block_import::ImportBlock

pub struct ImportBlock<Block: BlockT> {
    pub origin: BlockOrigin,
    pub header: Block::Header,
    pub justification: Option<Justification>,
    pub post_digests: Vec<DigestItemFor<Block>>,
    pub body: Option<Vec<Block::Extrinsic>>,
    pub finalized: bool,
    pub auxiliary: Vec<(Vec<u8>, Option<Vec<u8>>)>,
    pub fork_choice: ForkChoiceStrategy,
}

Data required to import a Block

Fields
origin: BlockOrigin

Origin of the Block

header: Block::Header

The header, without consensus post-digests applied. This should be in the same state as it comes out of the runtime.

Consensus engines which alter the header (by adding post-runtime digests) should strip those off in the initial verification process and pass them via the post_digests field. During block authorship, they should not be pushed to the header directly.

The reason for this distinction is so the header can be directly re-executed in a runtime that checks digest equivalence -- the post-runtime digests are pushed back on after.

justification: Option<Justification>

Justification provided for this block from the outside.

post_digests: Vec<DigestItemFor<Block>>

Digest items that have been added after the runtime for external work, like a consensus signature.

body: Option<Vec<Block::Extrinsic>>

Block's body

finalized: bool

Is this block finalized already? true implies instant finality.

auxiliary: Vec<(Vec<u8>, Option<Vec<u8>>)>

Auxiliary consensus data produced by the block. Contains a list of key-value pairs. If values are None, the keys will be deleted.

fork_choice: ForkChoiceStrategy

Fork choice strategy of this import.

Methods

impl<Block: BlockT> ImportBlock<Block>

pub fn into_inner(
self
) -> (BlockOrigin, <Block as BlockT>::Header, Option<Justification>, Vec<DigestItemFor<Block>>, Option<Vec<<Block as BlockT>::Extrinsic>>, bool, Vec<(Vec<u8>, Option<Vec<u8>>)>)

Deconstruct the justified header into parts.

pub fn post_header(&self) -> Cow<Block::Header>

Get a handle to full header (with post-digests applied).