Trait substrate_consensus_common::block_import::JustificationImport

pub trait JustificationImport<B: BlockT> {
    type Error: Error + Send + 'static;
    fn import_justification(
        &self, 
        hash: B::Hash, 
        number: NumberFor<B>, 
        justification: Justification
    ) -> Result<(), Self::Error>;

    fn on_start(&self, _link: &dyn Link<B>) { ... }
}

Justification import trait

Associated Types

type Error: Error + Send + 'static

Loading content...
Required methods

fn import_justification(
&self,
hash: B::Hash,
number: NumberFor<B>,
justification: Justification
) -> Result<(), Self::Error>

Import a Block justification and finalize the given block.

Loading content...
Provided methods

fn on_start(&self, _link: &dyn Link<B>)

Called by the import queue when it is started.

Loading content...