Module substrate_consensus_common::import_queue

Import Queue primitive: something which can verify and import blocks.

This serves as an intermediate and abstracted step between synchronization and import. Each mode of consensus will have its own requirements for block verification. Some algorithms can verify in parallel, while others only sequentially.

The ImportQueue trait allows such verification strategies to be instantiated. The BasicQueue and BasicVerifier traits allow serial queues to be instantiated simply.

Structs

BasicQueue
Interface to a basic block import queue that is importing blocks sequentially in a separate thread, with pluggable verification.

BasicSyncQueue
Basic block import queue that performs import in the caller thread.

IncomingBlock
Block data used by the queue.

Enums

BlockImportError
Block import error.

BlockImportMsg

BlockImportResult
Block import successful result.

BlockImportWorkerMsg

Traits

ImportQueue
Blocks import queue API.

Link
Hooks that the verification queue can use to influence the synchronization algorithm.

Verifier
Verify a justification of a block

Functions

import_single_block
Single block import function.

Type Definitions

Origin
Maps to the Origin used by the network.

SharedBlockImport
Shared block import struct used by the queue.

SharedFinalityProofImport
Shared finality proof import struct used by the queue.

SharedFinalityProofRequestBuilder
Shared finality proof request builder struct used by the queue.

SharedJustificationImport
Shared justification import struct used by the queue.