Trait substrate_consensus_common::import_queue::Link

pub trait Link<B: BlockT>: Send {
    fn blocks_processed(
        &mut self, 
        _imported: usize, 
        _count: usize, 
        _results: Vec<(Result<BlockImportResult<NumberFor<B>>, BlockImportError>, B::Hash)>
    ) { ... }

    fn justification_imported(
        &mut self, 
        _who: Origin, 
        _hash: &B::Hash, 
        _number: NumberFor<B>, 
        _success: bool
    ) { ... }

    fn request_justification(&mut self, _hash: &B::Hash, _number: NumberFor<B>) { ... }

    fn finality_proof_imported(
        &mut self, 
        _who: Origin, 
        _request_block: (B::Hash, NumberFor<B>), 
        _finalization_result: Result<(B::Hash, NumberFor<B>), ()>
    ) { ... }

    fn request_finality_proof(&mut self, _hash: &B::Hash, _number: NumberFor<B>) { ... }
}

Hooks that the verification queue can use to influence the synchronization algorithm.

Provided methods

fn blocks_processed(
&mut self,
_imported: usize,
_count: usize,
_results: Vec<(Result<BlockImportResult<NumberFor<B>>, BlockImportError>, B::Hash)>
)

Batch of blocks imported, with or without error.

fn justification_imported(
&mut self,
_who: Origin,
_hash: &B::Hash,
_number: NumberFor<B>,
_success: bool
)

Justification import result.

fn request_justification(&mut self, _hash: &B::Hash, _number: NumberFor<B>)

Request a justification for the given block.

fn finality_proof_imported(
&mut self,
_who: Origin,
_request_block: (B::Hash, NumberFor<B>),
_finalization_result: Result<(B::Hash, NumberFor<B>), ()>
)

Finality proof import result.

Even though we have asked for finality proof of block A, provider could return proof of some earlier block B, if the proof for A was too large. The sync module should continue asking for proof of A in this case.

fn request_finality_proof(&mut self, _hash: &B::Hash, _number: NumberFor<B>)

Request a finality proof for the given block.

Loading content...