Crate substrate_consensus_common

Common utilities for building and using consensus engines in substrate.

Much of this crate is unstable and thus the API is likely to undergo change. Implementors of traits should not rely on the interfaces to remain the same.

Re-exports

pub use self::error::Error;

pub use self::error::ErrorKind;

Modules

error
Error types in Consensus

evaluation
Block evaluation and evaluation errors.

import_queue
Import Queue primitive: something which can verify and import blocks.

offline_tracker
Tracks offline validators.

well_known_cache_keys
A list of all well known keys in the cache.

Structs

ImportBlock
Data required to import a Block

ImportedAux
Auxiliary data associated with an imported block result.

InherentData
Inherent data to include in a block.

NoNetwork
A synchronization oracle for when there is no network.

Enums

BlockOrigin
Block data origin.

ForkChoiceStrategy
Fork choice strategy.

ImportResult
Block import result.

Traits

Authorities
Trait for getting the authorities at a given block.

BlockImport
Block import trait.

Environment
Environment producer for a Consensus instance. Creates proposer instance and communication streams.

JustificationImport
Justification import trait

Proposer
Logic for a proposer.

SyncOracle
An oracle for when major synchronization work is being undertaken.